Cleaned up code and seperated into modules. Also improved performance with collision updates, by changing a ++ into a :
